feat: autentica as rotas de perfil

+++ app/Repositorios/PerfilRepositorio.cs
@@ -1,6 +1,4 @@
-using System.Runtime.CompilerServices;
 using api;
-using api.Perfis;
 using app.Entidades;
 using app.Repositorios.Interfaces;
 using Microsoft.EntityFrameworkCore;
@@ -41,6 +39,11 @@ namespace app.Repositorios
 
         public void RemovePerfil(Perfil perfil)
         {
+            if(perfil.Tipo == TipoPerfil.Basico || perfil.Tipo == TipoPerfil.Administrador)
+            {
+                throw new InvalidOperationException("Esse Perfil não pode ser excluido.");
+            }
+
             dbContext.Perfis.Remove(perfil);
         }
